requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
Django==3.0.8
django-easy-pdf==0.1.1
django-tables2==2.3.1
django-utils-six==2.0
html5lib==1.1
Pillow==7.2.0
psycopg2==2.8.5
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.44
six==1.15.0
sqlparse==0.3.1
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.4
